refactor: component name change
