+++ src/pages/group/MyGroup.tsx
@@ -1,0 +1,5 @@
+const MyGroup = () => {
+  return <div>MyGroup</div>;
+};
+
+export default MyGroup;
